Merge : 프로젝트에 토큰 등록, 수정, 해제 기능

See merge request s07-final/S07P31B207!40

+++ chilisource/microservices/project-service/src/main/java/com/ssafy/service/ProjectServiceImpl.java
@@ -1,10 +1,15 @@
 package com.ssafy.service;
 
+import com.ssafy.client.AuthServiceClient;
+import com.ssafy.config.loginuser.User;
 import com.ssafy.dto.request.ProjectCreateRequest;
+import com.ssafy.dto.request.ProjectTokenUpdateRequest;
 import com.ssafy.dto.request.ProjectUpdateRequest;
 import com.ssafy.dto.response.ProjectResponse;
+import com.ssafy.dto.response.TokenResponse;
 import com.ssafy.entity.Project;
 import com.ssafy.entity.UserProject;
+import com.ssafy.exception.NotAuthorizedException;
 import com.ssafy.exception.NotFoundException;
 import com.ssafy.repository.ProjectRepo;
 import com.ssafy.repository.RoleRepo;
@@ -13,10 +18,13 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import java.util.ArrayList;
 import java.util.List;
 import java.util.stream.Collectors;
 
-import static com.ssafy.exception.NotFoundException.PROJECT_NOT_FOUND;
+import static com.ssafy.exception.NotAuthorizedException.CREATE_NOT_AUTHORIZED;
+import static com.ssafy.exception.NotAuthorizedException.REMOVE_NOT_AUTHORIZED;
+import static com.ssafy.exception.NotFoundException.*;
 
 @RequiredArgsConstructor
 @Service
@@ -25,17 +33,21 @@ public class ProjectServiceImpl implements ProjectService {
     private final ProjectRepo projectRepo;
     private final UserProjectRepo userProjectRepo;
     private final RoleRepo roleRepo;
+    private final AuthServiceClient authServiceClient;
 
     @Override
     public ProjectResponse getProject(Long projectId) {
         Project project = projectRepo.findById(projectId)
                 .orElseThrow(() -> new NotFoundException(PROJECT_NOT_FOUND));
+
         return ProjectResponse.builder()
                 .id(project.getId())
                 .name(project.getName())
-                .teamName(project.getTeamName())
+                .description(project.getDescription())
                 .image(project.getImage())
                 .jiraProject(project.getJiraProject())
+                .gitRepo(project.getGitRepo())
+                .tokenList(getTokenList(project))
                 .build();
     }
 
@@ -48,9 +60,11 @@ public class ProjectServiceImpl implements ProjectService {
                 .map(project -> ProjectResponse.builder()
                         .id(project.getId())
                         .name(project.getName())
-                        .teamName(project.getTeamName())
+                        .description(project.getDescription())
                         .image(project.getImage())
                         .jiraProject(project.getJiraProject())
+                        .gitRepo(project.getGitRepo())
+                        .tokenList(getTokenList(project))
                         .build())
                 .collect(Collectors.toList());
     }
@@ -61,7 +75,7 @@ public class ProjectServiceImpl implements ProjectService {
     public void createProject(ProjectCreateRequest request, Long userId) {
         Project project = Project.builder()
                 .name(request.getName())
-                .teamName(request.getName())
+                .description(request.getDescription())
                 .image(request.getImage())
                 .build();
         projectRepo.save(project);
@@ -82,18 +96,101 @@ public class ProjectServiceImpl implements ProjectService {
     public void updateProject(ProjectUpdateRequest request) {
         Project project = projectRepo.findById(request.getId())
                 .orElseThrow(() -> new NotFoundException(PROJECT_NOT_FOUND));
-        project.update(request.getName(), request.getTeamName(), request.getImage(), request.getJiraProject(), null, null);
+        project.update(request.getName(), request.getDescription(), request.getImage());
     }
 
     // 프로젝트 삭제
     @Override
     @Transactional
     public void deleteProject(Long projectId, Long userId) {
-        // 프로젝트 삭제 권한 체크
+        UserProject userProjectManager = userProjectRepo.findByUserIdAndProjectId(userId, projectId)
+                .orElseThrow(() -> new NotFoundException(USER_PROJECT_NOT_FOUND));
+        if (userProjectManager.getRole().getRemove()) {
+            Project project = projectRepo.findById(projectId)
+                    .orElseThrow(() -> new NotFoundException(PROJECT_NOT_FOUND));
+            projectRepo.delete(project);
+        } else {
+            throw new NotAuthorizedException(REMOVE_NOT_AUTHORIZED);
+        }
+    }
 
-        // 삭제
+    @Override
+    public void updateProjectToken(User user, ProjectTokenUpdateRequest request) {
+        Project project = projectRepo.findById(request.getProjectId())
+                .orElseThrow(() -> new NotFoundException(PROJECT_NOT_FOUND));
+
+        UserProject userProjectManager = userProjectRepo.findByUserIdAndProjectId(user.getId(), request.getProjectId())
+                .orElseThrow(() -> new NotFoundException(USER_PROJECT_NOT_FOUND));
+
+        if (!userProjectManager.getRole().getName().equalsIgnoreCase("MASTER")) {
+            throw new NotAuthorizedException(CREATE_NOT_AUTHORIZED);
+        }
+
+        List<TokenResponse> tokenResponses = authServiceClient.getToken(user);
+
+        TokenResponse tokenResponse = null;
+        switch (request.getName().toUpperCase()) {
+            case "JIRA":
+                for (TokenResponse tr : tokenResponses) {
+                    if (tr.getTokenCodeId() == 0L) {
+                        tokenResponse = tr;
+                    }
+                }
+
+                if (tokenResponse != null) {
+                    project.updateJira(tokenResponse.getValue(), request.getDetail());
+                } else {
+                    throw new NotFoundException(TOKEN_NOT_CONNECTED);
+                }
+                break;
+            case "GIT":
+                for (TokenResponse tr : tokenResponses) {
+                    if (tr.getTokenCodeId() == 1L) {
+                        tokenResponse = tr;
+                    }
+                }
+
+                if (tokenResponse != null) {
+                    project.updateGit(tokenResponse.getValue(), request.getDetail());
+                } else {
+                    throw new NotFoundException(TOKEN_NOT_CONNECTED);
+                }
+                break;
+            default:
+                break;
+        }
+    }
+
+    @Override
+    public void deleteProjectToken(User user, Long projectId, String name) {
         Project project = projectRepo.findById(projectId)
                 .orElseThrow(() -> new NotFoundException(PROJECT_NOT_FOUND));
-        projectRepo.delete(project);
+
+        UserProject userProjectManager = userProjectRepo.findByUserIdAndProjectId(user.getId(), projectId)
+                .orElseThrow(() -> new NotFoundException(USER_PROJECT_NOT_FOUND));
+
+        if (!userProjectManager.getRole().getName().equalsIgnoreCase("MASTER")) {
+            throw new NotAuthorizedException(REMOVE_NOT_AUTHORIZED);
+        }
+
+        switch (name.toUpperCase()) {
+            case "JIRA":
+                project.deleteJira();
+                break;
+            case "GIT":
+                project.deleteGit();
+                break;
+            default:
+                break;
+        }
+    }
+
+    private List<String> getTokenList(Project project) {
+        List<String> tokenList = new ArrayList<>();
+
+        if (project.getJiraToken() != null) tokenList.add("JIRA");
+        if (project.getGitToken() != null) tokenList.add("GIT");
+
+        return tokenList;
     }
 }
